requirements.txt
future
emoji
requests
sqlalchemy==1.3.23
python-telegram-bot==11.1.0
psycopg2-binary
feedparser
pyowm==2.8.0
pythonping
hurry.filesize
